% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{annotate cleaned dataset}
\usage{
annotate(.data)
}
\arguments{
\item{.data}{cleaned dataset to annotate}
}
\value{
annotated dataset
}
\description{
annotate cleaned dataset
}
\examples{
export2dataframe('data.txt') \%>\% annotate()
}
