% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{AppendClass}
\alias{AppendClass}
\title{Append new class}
\usage{
AppendClass(obj, labels)
}
\arguments{
\item{obj}{object}

\item{labels}{array of class names to append}
}
\value{
object with new class atrribute
}
\description{
Append new class
}
\examples{
x <- c('A', 'B')
AppendClass(x, 'LETTER')
}
