% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{Boltzmann}
\alias{Boltzmann}
\title{Boltzmann model for fitting time series data}
\usage{
Boltzmann(time_, val_, A0 = 1, k0 = 1, t20 = 1)
}
\arguments{
\item{time_}{time series}

\item{val_}{normalized value}

\item{A0}{amplititude}

\item{k0}{rate constant}

\item{t20}{halt time}
}
\value{
a model
}
\description{
Boltzmann model for fitting time series data
}
\examples{
Boltzmann(1:10, c(0,0,1,3,5,7,9,10,10,10), A0 = 10, k0 = 10, t20 = 5)

}
