% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ThT.R
\name{Boltzmann_double}
\alias{Boltzmann_double}
\title{Boltzmann model for fitting time series data}
\usage{
Boltzmann_double(time_, val_, start = list(y0 = 500, A = 500, k = 1.1, t2 =
  10, A2 = 500, k2 = 1.1, t22 = 10))
}
\arguments{
\item{start}{}
}
\description{
Boltzmann model for fitting time series data
}
