% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{export2dataframe}
\alias{export2dataframe}
\title{export2dataframe
return cleaned data.frame from Gene5 exported tab-delim data}
\usage{
export2dataframe(filename, Ctrl = list(sample.by = "row"))
}
\arguments{
\item{Ctrl}{}
}
\value{
cleaned dataset
}
\description{
export2dataframe
return cleaned data.frame from Gene5 exported tab-delim data
}
\examples{

}
