% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{fill_str_array}
\alias{fill_str_array}
\title{fill empty string in an array using the non-empty occurance.}
\usage{
fill_str_array(vec)
}
\arguments{
\item{vec}{character array}
}
\description{
fill empty string in an array using the non-empty occurance.
}
\examples{
fill_str_array(c("A", "", "B", "", "", "C"))
fill_str_array(c("A", "", "B", NA, "", "C"))
fill_str_array(c("", "B", "", "", "C"))

}
