% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{filter.fit}
\alias{filter.fit}
\title{Clean deviated fit}
\usage{
filter.fit(df, TOL = 1)
}
\arguments{
\item{df}{NotImplemented}

\item{TOL}{NotImplemented}
}
\value{
NotImplemented
}
\description{
remove data points of which standar deviation of fitting parameter is larger
than TOL
}
\examples{
NotImplemented
}
