% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ThT.R
\name{fit.boltzmann.double}
\alias{fit.boltzmann.double}
\title{Title}
\usage{
fit.boltzmann.double(ds, start = list(A = 3000, y0 = 1000, k = 10, t2 = 1, A2
  = 3000, k2 = 10, t22 = 1))
}
\arguments{
\item{start}{}
}
\description{
Title
}
