% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ThT.R
\name{fit.boltzmann}
\alias{fit.boltzmann}
\title{Title}
\usage{
fit.boltzmann(ds, start = list(A = 3000, y0 = 1000, k = 10, t2 = 1))
}
\arguments{
\item{ds}{gen5helper annotated data.frame}

\item{start}{a list of initial fitting para}
}
\value{
ds with extra fitting features
}
\description{
Title
}
