% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp}
\alias{fp}
\title{find sigmoid curve peak
DEPRECATED! use Boltzmann() instead}
\usage{
fp(vec, x)
}
\arguments{
\item{x}{}
}
\description{
find sigmoid curve peak
DEPRECATED! use Boltzmann() instead
}
