% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{g5h.clean}
\alias{g5h.clean}
\title{g5h.clean}
\usage{
g5h.clean(filename, Ctrl = list(sample.by = "row"))
}
\arguments{
\item{filename}{character}

\item{Ctrl}{list}
}
\value{
data.frame
}
\description{
Return technically correct data.frame from Gen5 2.06 exported tab-delim data
}
\examples{
g5h.clean('data.txt')
}
