% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{g5h.gather_col}
\alias{g5h.gather_col}
\title{Add mean and standard deviation}
\usage{
g5h.gather_col(.data)
}
\arguments{
\item{.data}{data.frame}
}
\value{
data.frame appended with val.m and val.sd
}
\description{
g5h.gather_col() preserve existing variables and add mean and standard
deviation, grouped by col.
}
\details{
g5h.gather_row() preserve existing variables and add mean and standard
deviation, grouped by row.
}
\examples{
# group by col
g5h.clean('data/demo.txt') \%>\%
    g5h.gather_col()
# group by row
g5h.clean('data/demo.txt') \%>\%
    g5h.gather_row()
}
