% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{g5h.map_row}
\alias{g5h.map_row}
\title{Map row or col to new values}
\usage{
g5h.map_row(.data, feature, factors)
}
\arguments{
\item{.data}{data.frame cleaned by g5h.clean()}

\item{feature}{character, name of new variable to add.}

\item{factors}{vector with length equal to levels of row or col.}
}
\value{
data.frame appended with new variables
}
\description{
g5h.map_row() or g5h.map_col() add new variables by mapping row or col to
new values.
}
\examples{
g5h.clean('data.txt') \%>\%
    g5h.map_row(feature = 'dose', factors = c(1, 2, 3)) \%>\%
    g5h.map_col(feature = 'treatment', factors = c('freeze', 'bake', 'grill'))
}
