% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{g5h.set_time2}
\alias{g5h.set_time2}
\title{Add time intervals to cleaned dataset}
\usage{
g5h.set_time2(.data, units = "hours")
}
\arguments{
\item{.data}{data.frame cleaned by g5h.clean()}

\item{units}{"hours" or "minutes"}
}
\value{
input data.frame appended with time
}
\description{
g5h.set_time() preserves existing variables and add new variable,
time, which are time intervals in hours.
}
