% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{gather_row}
\alias{gather_row}
\title{Add mean and standard deviation}
\usage{
gather_row(.data)
}
\arguments{
\item{.data}{data.frame}
}
\value{
data.frame appended with val.m and val.sd
}
\description{
gather_col() preserve existing variables and add mean and standard
deviation, grouped by col.
gather_row() preserve existing variables and add mean and standard
deviation, grouped by row.
}
