% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen5helper-package.r
\docType{package}
\name{gen5helper}
\alias{gen5helper}
\alias{gen5helper-package}
\title{gen5helper: A Collection of Functions for Processing Gen5 2.06 Exported Data}
\description{
A collection of functions for processing Gen5 2.06 exported data.
Gen 5 is a popular data analysis software for BioTek plate readers. This
packages contains functions for data cleaning, modeling and plotting using
exported data from Gen5 version 2.06. It exports technically correct data
defined in (Edwin de Jonge and Mark van der Loo, 2013) for customized
analysis. It contains Boltzmann fitting for general kinetic analysis. It
also implement line plot and bar plot for generating publishable figures.
See https://www.github.com/yanxianUCSB/gen5helper for more information,
documentation and examples.
}
