% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{is_spike}
\alias{is_spike}
\title{this is a file for functions that are universally useful at common data manipulations
Is a Value Spike?}
\usage{
is_spike(x, tol = 1)
}
\arguments{
\item{x}{numeric vector}

\item{tol}{tolerance}
}
\value{
boolean vector
}
\description{
this is a file for functions that are universally useful at common data manipulations
Is a Value Spike?
}
\examples{
is_spike(c(1, 2, 1, 1000, -100, 2, 1))
}
