% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{mapvalues_}
\alias{mapvalues_}
\title{Map the unique values of a vector}
\usage{
mapvalues_(x, facs, bNaturalSort = FALSE)
}
\arguments{
\item{x}{factor or character}

\item{facs}{character. It maps unique(x) to facs}

\item{bNaturalSort}{binary. Whether to convert factor in natural order.}
}
\value{
factor
}
\description{
Map the unique values of a vector
}
\examples{
mapvalues_(c("A","A","B","C"), c("one", "two", "three"))
mapvalues_(c("apple", "apple", "banana", "pineable", "pineable"), c(3, 2, 1), bNaturalSort = TRUE)

}
