% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{most.freq}
\alias{most.freq}
\title{Return most frequent numbers}
\usage{
most.freq(x, n = 1)
}
\arguments{
\item{x}{an array of elements}

\item{n}{integer, default is 1}
}
\value{
the most n elements
}
\description{
get the n most frequent elements in an array
}
\examples{
most.freq(c('a', 'a', 'b', 'b', 'b', 'c'), n = 2)
most.freq(c(1, 1, 2, 3, 3, 3, 4, 4), n = 2)

}
