% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{mutate_cond}
\alias{mutate_cond}
\title{this is a file for functions that are universally useful at common data manipulations
Create or transform variables conditionally}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{A tbl.}

\item{condition}{condition.}

\item{...}{Name-value pairs of expressions.}

\item{envir}{Environment.}
}
\description{
this is a file for functions that are universally useful at common data manipulations
Create or transform variables conditionally
}
\examples{
mtcars \%>\% mutate_cond(gear == 4, hp = 1000)

}
