% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{normalize}
\alias{normalize}
\title{Scale a vector to 0-1 by min and max}
\usage{
normalize(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric}

\item{na.rm}{bool whether to remove NA values.}
}
\value{
a normalized vector
}
\description{
Scale a vector to 0-1 by min and max
}
\examples{
normalize(0:10)
normalize(c(1, 100, NA, 10), na.rm = TRUE)
}
