% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.annotated.dataset}
\alias{plot.annotated.dataset}
\title{plot annotated dataset}
\usage{
\method{plot}{annotated.dataset}(.data, primary = "tht", secondary = "abs",
  i = NULL, facet.wrap = FALSE)
}
\arguments{
\item{.data}{version XXX}

\item{primary}{primary readingType default 'tht'}

\item{secondary}{secondary readingType default 'abs'}

\item{i}{sufix of images output filename}
}
\value{
.data
}
\description{
plot annotated dataset
}
\examples{
export2data.frame('data/demo.txt') \%>\%
    annotate() \%>\%
    plot.annotated.dataset()
}
