% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fitted.dataset}
\alias{plot.fitted.dataset}
\title{plot fitted dataset}
\usage{
\method{plot}{fitted.dataset}(df, tiff = F, i = NULL)
}
\arguments{
\item{df}{version XXX}

\item{tiff}{logical, whether to output tiff imgs}

\item{i}{suflix of images output filename}
}
\value{
df
}
\description{
plot fitted dataset
}
\examples{
g5h.clean('data/demo.txt') \%>\%
    g5h.annotate() \%>\%
    fit.Boltzmann() \%>\%
    plot.fitted.dataset()
}
