% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotFlsAbs}
\alias{plotFlsAbs}
\title{Plot multiple readingType in line+ribbon,
the linewidth of the first reading Type will be doubled}
\usage{
plotFlsAbs(ds, Ctrl = list(flsReadingType = unique(ds$readingType)[1], plotN =
  c("all", "major")[2]))
}
\arguments{
\item{ds}{a gen5 annotated data.frame}
}
\value{
a ggplot2 object
}
\description{
Plot multiple readingType in line+ribbon,
the linewidth of the first reading Type will be doubled
}
\examples{
g5h.clean('demo/data.txt') \%>\%
    g5h.annotate() \%>\%
    plotFlsAbs(Ctrl = list(flsReadingType='tht', plotN='major'))
}
