% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{range_}
\alias{range_}
\title{Return range of a vector}
\usage{
range_(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric}

\item{na.rm}{bool whether to remove NA values.}
}
\value{
numeric value
}
\description{
This computes the range of a vector as a value.
}
\examples{
range_(c(1, 5, 10))
range_(c(1, 5, 10, NA), na.rm = TRUE)
}
