% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{ReplaceCond}
\alias{ReplaceCond}
\title{Replace Values in Multiple Variables Conditionally}
\usage{
ReplaceCond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{A tbl.}

\item{condition}{condition.}

\item{...}{Name-value pairs of expressions.}

\item{envir}{Environment.}
}
\description{
Replace Values in Multiple Variables Conditionally
}
\examples{
mtcars \%>\%
    ReplaceCond(gear == 4, hp = 1000, cyl = NA)

}
