% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{saveRDS_}
\alias{saveRDS_}
\title{saveRDS and return .data}
\usage{
saveRDS_(.data, file = file, ...)
}
\arguments{
\item{file}{filename}

\item{...}{for saveRDS}
}
\value{
.data
}
\description{
saveRDS and return .data
}
