% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{scale_head_tail}
\alias{scale_head_tail}
\title{Scale a Numeric Vector Using the Head and Tail}
\usage{
scale_head_tail(x, nhead, ntail)
}
\arguments{
\item{x}{a numeric vector}

\item{nhead}{the first nhead of x will be averaged and scaled to 0}

\item{ntail}{the last ntail of x will be averaged and scaled to 1}
}
\description{
Scale a Numeric Vector Using the Head and Tail
}
\examples{
scale_head_tail(1:10, nhead = 2, ntail = 2)

}
