% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{scale_quantile}
\alias{scale_quantile}
\title{this is a file for functions that are universally useful at common data manipulations
Use Sample Quantiles to Scale a Numeric Vector}
\usage{
scale_quantile(x, qmin, qmax)
}
\arguments{
\item{x}{a numeric vector}

\item{qmin}{x which below qmin will be averaged and scaled to 0}

\item{qmax}{x which above qmax will be averaged and scaled to 1}
}
\description{
this is a file for functions that are universally useful at common data manipulations
Use Sample Quantiles to Scale a Numeric Vector
}
\examples{
scale_quantile(c(1:10), 0.25, 0.75)
scale_quantile(c(1:10, NA, 10:1), 0.25, 0.75)

}
