% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{select.reading}
\alias{select.reading}
\title{filter matching reading type}
\usage{
select.reading(.data, rt)
}
\arguments{
\item{.data}{data.frame cleaned by g5h.clean()}

\item{rt}{readingType}
}
\description{
filter matching reading type
}
