% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{select.wells}
\alias{select.wells}
\title{filter matching wells}
\usage{
\method{select}{wells}(.data, rows, cols)
}
\arguments{
\item{.data}{data.frame cleaned by g5h.clean()}

\item{rows}{rows to be selected}

\item{cols}{columns to be selected}
}
\description{
filter matching wells
}
