% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{SelectWells}
\alias{SelectWells}
\title{Select Wells}
\usage{
SelectWells(.data, rows, cols)
}
\arguments{
\item{.data}{tbl g5h.cleaned}

\item{rows}{character array of rows}

\item{cols}{array of cols}
}
\value{
tbl subsetted
}
\description{
Select Wells
}
\examples{
\donttest{
# suppose "gen5_export.txt" is the export from Gen5 2.06

g5h.clean2("gen5_export.txt") \%>\%
    g5h.annotate() \%>\%
    SelectWells(c('A', 'C'), 2:4)
}
}
