% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{seq.LETTERS}
\alias{seq.LETTERS}
\title{this is a file for functions that are universally useful at common data manipulations
seq.LETTERS
generate a sequence of LETTERS from FROM to TO}
\usage{
\method{seq}{LETTERS}(from, to)
}
\arguments{
\item{from}{LETTER}

\item{to}{LETTER}
}
\description{
this is a file for functions that are universally useful at common data manipulations
seq.LETTERS
generate a sequence of LETTERS from FROM to TO
}
