% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{seq_letters}
\alias{seq_letters}
\title{Generate a sequence of lower-case Roman alphabet}
\usage{
seq_letters(from, to)
}
\arguments{
\item{from}{first letter}

\item{to}{last letter}
}
\value{
an array of letters
}
\description{
Generate a sequence of lower-case Roman alphabet
}
\examples{
seq_letters('b', 'e')

}
