% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{smooth.mean}
\alias{smooth.mean}
\title{Smooth a vector using moving average}
\usage{
smooth.mean(vec, naverage)
}
\arguments{
\item{vec}{numeric vector}

\item{naverage}{width of moving average}
}
\value{
smoothed
}
\description{
Smooth a vector using moving average
}
\examples{
smooth.mean(1:10, 2)
smooth.mean(1:10, 3)
smooth.mean(1:10, 5)
}
