% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ThT.R
\name{ThT_fit}
\alias{ThT_fit}
\title{Fitting ThT time series data with Boltzmann model. return fit data and model}
\usage{
ThT_fit(time_, val_)
}
\arguments{
\item{val_}{}
}
\value{
list(data.frame, model)
}
\description{
Fitting ThT time series data with Boltzmann model. return fit data and model
}
