% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ThT.R
\name{ThT_fit_ThT}
\alias{ThT_fit_ThT}
\title{ThT_fit function to get only fit data.}
\usage{
ThT_fit_ThT(time_, val_)
}
\arguments{
\item{val_}{}
}
\value{
numeric vector of fit data
}
\description{
ThT_fit function to get only fit data.
}
