% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{ui.clean}
\alias{ui.clean}
\title{UI Clean
User interface for cleaning Gen5 automatic exported .TXT data}
\usage{
ui.clean(args = commandArgs(trailingOnly = T))
}
\arguments{
\item{args}{command arguments}
}
\value{
cleaned dataset
}
\description{
UI Clean
User interface for cleaning Gen5 automatic exported .TXT data
}
