% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ThT.R
\name{ui.plot.fit}
\alias{ui.plot.fit}
\title{UI plot fit
User interface for ploting fitted ThT kinetic}
\usage{
ui.plot.fit()
}
\description{
UI plot fit
User interface for ploting fitted ThT kinetic
}
