% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{ungroup_}
\alias{ungroup_}
\title{Run ungroup() and as.data.frame()}
\usage{
ungroup_(.data)
}
\arguments{
\item{.data}{grouped data.frame}
}
\value{
an ungrouped data.frame()
}
\description{
Run ungroup() and as.data.frame()
}
\examples{
\donttest{
data <- data.frame(m=c(1,2), n=c(2,3), group=c('a','b'))
data \%>\%
    group_by(group) \%>\%
    ungroup_()
}
}
