% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{write.csv_}
\alias{write.csv_}
\title{write.csv without row.names}
\usage{
write.csv_(x, file)
}
\arguments{
\item{x}{object}

\item{file}{filename for write.csv}
}
\value{
whatever object passed in
}
\description{
This function returns the object passed in, can be used for dplyr pipeline.
}
\examples{
\donttest{
write.csv_(data.frame(a=1, b=2, c=3), file.path(tempdir(), "data.csv"))

data <- data.frame(a=1, b=2, c=3)
data \%>\%
    file.path(tempdir(), "data.csv") \%>\%
    print()
}
}
