% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{annotate
annotate cleaned data.frame}
\usage{
annotate(ds)
}
\arguments{
\item{ds}{}
}
\description{
annotate
annotate cleaned data.frame
}
\examples{
ds.cleaned   <- export2dataframe('data.txt')
ds.annotated <- annotate(ds.cleaned)
}
