% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{as.is}
\alias{as.is}
\title{this is a file for functions that are universally useful at common data manipulations
as.is}
\usage{
as.is(x, vec)
}
\arguments{
\item{x}{object to transform}

\item{vec}{object to extract class}
}
\description{
this is a file for functions that are universally useful at common data manipulations
as.is
}
