% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{Boltzmann}
\alias{Boltzmann}
\title{Boltzmann model for fitting time series data}
\usage{
Boltzmann(time_, val_, start = list(y0 = 500, A = 500, k = 1.1, t2 = 10))
}
\arguments{
\item{time_}{NotExported}

\item{val_}{NotExported}

\item{start}{start = list(y0 = 500, A = 500, k = 1.1, t2 = 10)}
}
\value{
NotExported
}
\description{
Boltzmann model for fitting time series data
}
