% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{factor2num}
\alias{factor2num}
\title{Convert factor to numeric}
\usage{
factor2num(x)
}
\arguments{
\item{x}{factor}
}
\description{
Convert factor to numeric
}
\examples{
factor2num(factor(c('1', '10', '100')))

}
