% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.boltzmann}
\alias{fit.boltzmann}
\title{Fit readings with Boltzmann model}
\usage{
fit.boltzmann(.data, start = list(A = 3000, y0 = 1000, k = 10, t2 = 1))
}
\arguments{
\item{.data}{data.frame output by g5h.clean()}

\item{start}{a list of initial guesses of fitting parameter}
}
\value{
data.frame with fitted parameters
}
\description{
fit.boltzmann() using Boltzmann model to fit readings and time intervals with
unit of hours, using start as initial guesses. It appends A, y0, k, t2 and val.predict, while preserving existing
variables.
}
\examples{
g5h.clean('data/demo.txt') \%>\%
    g5h.annotate() \%>\%
    fit.boltzmann(start = list(A = 3000, y0 = 1000, k = 10, t2 = 1))

}
