% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{g5h.annotate}
\alias{g5h.annotate}
\title{Add useful variables}
\usage{
g5h.annotate(.data, by = "col")
}
\arguments{
\item{.data}{data.frame cleaned by g5h.clean()}

\item{by}{'col' or 'row', default is 'col'. See ?g5h.gather_col for more info.}
}
\value{
data.frame
}
\description{
Add time interval, mean, standard deviation and initilized treatment and dose.
}
\examples{
ds <- g5h.clean('data/demo.txt')
ds \%>\% g5h.annotate()
ds \%>\% g5h.annotate(by='row')

}
