% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{g5h.clean2}
\alias{g5h.clean2}
\title{Clean Gen5 exported data}
\usage{
g5h.clean2(files)
}
\arguments{
\item{files}{a vector of names of the file which the data are to be read from.
If it does not contain an absolute path, the file name is relative to the
current working directory, getwd().}
}
\value{
technically correct data.frame.
}
\description{
g5h.clean2() returns technically correct data.frame from Gen5 2.06 exported
tab-delim data. The exported data can be generated using default export
protocol in Gen5 2.06. See Gen5 User Guide for more information.
}
