% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{g5h.clean_}
\alias{g5h.clean_}
\title{Clean Gen5 exported data}
\usage{
g5h.clean_(file)
}
\arguments{
\item{file}{the name of the file which the data are to be read from. If it
does not contain an absolute path, the file name is relative to the current
working directory, getwd().}
}
\value{
technically correct data.frame.
}
\description{
g5h.clean_() returns technically correct data.frame from Gen5 2.06 exported
tab-delim data. The exported data can be generated using default export
protocol in Gen5 2.06. See Gen5 User Guide for more information.
}
\examples{
\dontrun{
# suppose "exported_data.txt" is the exports from Gen5 2.06

# this line will clean one exported data
data <- g5h.clean2("exported_data.txt")
}
transfrom readLines() output to data.frame
}
