% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{mapvalues_}
\alias{mapvalues_}
\title{mapvalues_}
\usage{
mapvalues_(x, facs, bNaturalSort = F)
}
\arguments{
\item{x}{factor or character}

\item{facs}{character. It maps unique(x) to facs}

\item{bNaturalSort}{binary}
}
\value{
factor
}
\description{
mapvalues_
}
