% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{mapvalues_}
\alias{mapvalues_}
\title{mapvalues_
Map the unique values of a vector}
\usage{
mapvalues_(x, facs, bNaturalSort = F)
}
\arguments{
\item{x}{factor or character}

\item{facs}{character. It maps unique(x) to facs}

\item{bNaturalSort}{binary}
}
\value{
factor
}
\description{
mapvalues_
Map the unique values of a vector
}
\examples{
mapvalues_(c("A","A","B","C"), c("one", "two", "three"))
}
