% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{range_}
\alias{range_}
\title{Range of a vector}
\usage{
range_(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric value
}
\description{
This computes the range of a vector as a value; NA removed.
}
