% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{saveRDS_}
\alias{saveRDS_}
\title{saveRDS and return .data}
\usage{
saveRDS_(.data, file, ...)
}
\arguments{
\item{.data}{object to be saved}

\item{file}{filename to save}

\item{...}{for saveRDS}
}
\value{
whatever object passed in.
}
\description{
This function returns the object passed in, can be used for dplyr pipeline.
}
\examples{
\donttest{
data <- data.frame(a=1, b=2, c=3)
data \%>\%
    saveRDS_(file.path(tempdir(), "data.rds")) \%>\%
    print()
}

}
