% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{ui.clean}
\alias{ui.clean}
\title{UI for g5h.clean}
\usage{
ui.clean(args = commandArgs(trailingOnly = T))
}
\arguments{
\item{args}{command arguments}
}
\value{
data.frame
}
\description{
ui.clean() is a command line user interface for g5h.clean. See ?g5h.clean for
more information.
}
