% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ui.fit.ThT}
\alias{ui.fit.ThT}
\title{UI fit ThT}
\usage{
ui.fit.ThT(START = list(A = 3000, y0 = 1000, k = 1, t2 = 5))
}
\arguments{
\item{START}{initial guess for fit.Boltzmann}
}
\description{
User interface for fitting ThT kinetic curve
}
