% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{ungroup_}
\alias{ungroup_}
\title{Ungroup() and as.data.frame()}
\usage{
ungroup_(.data)
}
\arguments{
\item{.data}{grouped data.frame}
}
\value{
data.frame()
}
\description{
Ungroup() and as.data.frame()
}
