% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{write.csv_}
\alias{write.csv_}
\title{write.csv and return .data}
\usage{
write.csv_(x, file)
}
\arguments{
\item{x}{object}

\item{file}{filename for write.csv}
}
\value{
x
}
\description{
write.csv and return .data
}
